Bootstrap/Bootswatch code for the css
I added a banner to the homes screens and moved the login/signup to the right. The logged in as "user" is also moved to the left. The sign in and signup pages are also redesigned to be centered. They are on a white background for now but can be changed later if we do not like it. The spacing looks better and the input boxes also have a better look. At the top of the boxes they have a link to go back to the main page.